requirements.txt
# ─── Web framework ───────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9      # required for file uploads in FastAPI

# ─── Google Earth Engine ──────────────────────────────────────────────────────
earthengine-api==0.1.409

# ─── ML / DL ──────────────────────────────────────────────────────────────────
tensorflow==2.16.1            # includes Keras 3 & TFLite converter
numpy==1.26.4
Pillow==10.3.0                # image processing for inference

# ─── HTTP client (async weather API calls) ───────────────────────────────────
httpx==0.27.0

# ─── Config ───────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.7.1               # bundled with FastAPI, pinned for clarity

# ─── Optional: auth (JWT) ────────────────────────────────────────────────────
PyJWT==2.8.0

# ─── Optional: background tasks ──────────────────────────────────────────────
# celery==5.4.0
# redis==5.0.4

# ─── Dev / testing ────────────────────────────────────────────────────────────
pytest==8.2.2
httpx==0.27.0                 # also used by TestClient
